+++ crates/rbuilder/Cargo.toml
@@ -137,7 +137,6 @@ assert_matches = "1.5.0"
 criterion = { version = "0.5.1", features = ["html_reports", "async_tokio"] }
 
 [features]
-nethermind_evm = []
 # TODO: remove?
 optimism = []
 redact-sensitive = []
