+++ crates/rbuilder/src/building/evm/mod.rs
@@ -2,6 +2,7 @@ use alloy_evm::{eth::EthEvmContext, Database, Evm, EvmEnv, IntoTxEnv};
 use revm::{
     context::{result::ResultAndState, TxEnv},
     context_interface::result::{EVMError, HaltReason},
+    interpreter::interpreter::EthInterpreter,
     Inspector,
 };
 
@@ -12,37 +13,36 @@ pub trait BuilderEvm<DB: Database> {
     ) -> Result<ResultAndState<HaltReason>, EVMError<DB::Error>>;
 }
 
-impl<T, DB: Database> BuilderEvm<DB> for T
+impl<DB, EVM> BuilderEvm<DB> for EVM
 where
-    T: Evm<DB = DB, Tx = TxEnv, Error = EVMError<DB::Error>, HaltReason = HaltReason>,
+    DB: Database<Error: Send + Sync + 'static>,
+    EVM: Evm<DB = DB, Tx = TxEnv, Error = EVMError<DB::Error>, HaltReason = HaltReason>,
 {
     fn transact(
         &mut self,
         tx: impl IntoTxEnv<TxEnv>,
     ) -> Result<ResultAndState<HaltReason>, EVMError<DB::Error>> {
-        T::transact(self, tx)
+        EVM::transact(self, tx)
     }
 }
 
 /// Main trait to create instance of EVM.
-/// Allows to use different implementations of EVM with a simpler, more concrete interface than `reth_evm::EvmFactory`.
-/// A type responsible for creating instances of an ethereum virtual machine given a certain input.
+/// Allows to use implementations of EVM with a simpler, more concrete interface than `reth_evm::EvmFactory`.
 pub trait EvmFactory {
-    fn create_evm<DB: Database>(&self, db: DB, env: EvmEnv) -> impl BuilderEvm<DB>;
+    fn create_evm<DB>(&self, db: DB, env: EvmEnv) -> impl BuilderEvm<DB>
+    where
+        DB: Database<Error: Send + Sync + 'static>;
 
-    fn create_evm_with_inspector<DB: Database, I: Inspector<EthEvmContext<DB>>>(
+    fn create_evm_with_inspector<DB, I>(
         &self,
         db: DB,
         env: EvmEnv,
         inspector: I,
-    ) -> impl BuilderEvm<DB>;
+    ) -> impl BuilderEvm<DB>
+    where
+        DB: Database<Error: Send + Sync + 'static>,
+        I: Inspector<EthEvmContext<DB>, EthInterpreter>;
 }
 
-mod nethermind_evm;
 mod revm_evm;
-
-// Default to RevmEvmFactory
-#[cfg(not(feature = "nethermind_evm"))]
 pub type RBuilderEvm = revm_evm::RevmEvmFactory;
-#[cfg(feature = "nethermind_evm")]
-pub type RBuilderEvm = nethermind_evm::NethermindEvmFactory;
